loyalty_db/requirements.txt
yoyo-migrations==8.2.0
psycopg2-binary==2.9.6